feat: record failed synthetic transactions
Closes AC-886. If a synthetic transaction is verified to originate from a BVN, record it even if it fails.

Previously, CheckTx failed if transaction-specific verification failed. Thus, if a synthetic transaction failed for some reason, it was not recorded. If a synthetic transaction failure is not recorded, there is no way to differentiate between 1) the transaction has not been delivered, 2) the transaction was dropped in transit, and 3) the transaction failed.

With this change, a synthetic transaction (that was validly produced by a BVN) will always be recorded, so the client can verify its status even if it fails.