fix: type gen
Closes AC-974. The original issue was, alternative names are handled with a custom JSON un/marshaller, but having a field with an alternative name did not prompt the generator to create a custom un/marshaller. Thus, if a type had a field with an alternative name but no fields needed special handling in JSON, an un/marshaller would not be created.

Then I figured out that un/marshalling of embedded types was potentially broken: If A embeds B and B has a custom un/marshaller, B's custom un/marshaller is used for `var a A`. That is bad, because any other fields in A are ignored. To get around this, the generator creates a custom un/marshaller for A, in which all of the embedded fields are unpacked.


Former-commit-id: 95b73e73bca0411389b4a3f487fd8ea7c962e58c