feat: use new marshaller for generated types
Closes AC-1047.

- The `slice` field for type specs has been replaced with `repeatable: true`.
- The builtin type `chainSet` has been replaced with `type: chain, repeatable: true`.
- The Accumulate binary format/encoding has been redesigned:

Each field now has a field number which is added as a prefix to the field when it is encoded. When decoding, if a field number is not present, that means the field is not present. For example, if `Name string` is optional, then during encoding if `Name` has the default value it is omitted completely from the result.

Also, changes `Transaction.Body` from `[]byte` to `TransactionPayload`


Former-commit-id: 9297b66cb2291c6019cee081d0b0ecd11d3a03cd