feat(prot): add get range function to the BPT
FIX: FEAT: Fixed the persistence issues with BPT, Added a Get function to BPT to look up a key (for testing)

Closes AC-1145, AC-1119.

Some code around inserting new key/value pairs into the BPT existed around if the node at a power of two was in the persisted block or not.  This led to various errors when the assumption was the nde was in the persisted block.

Added a Get function to the BPT in order to make testing easier.