fix(prot): synthetic transactions always succeed
Closes AC-1176. Synthetic transactions must not fail. Specifically, if `CheckTx` returns a non-zero error code, the entire batch is rejected. If that batch happens to be synthetic transactions from another BVN, all synthetic transactions are rejected. If that batch happens to be from the local BVN to itself, it will also include internal transactions which will also be rejected. That will completely stall the synthetic transaction signing and sending process.

However, users must not be allowed to submit garbage without paying for it. This MR solves this conflict by returning a non-zero error code if and only if the batch includes user transactions. Since the batch used for synthetic transactions should never include a synthetic transaction, this solves the issue.


Former-commit-id: c6ac17e399a7e5c379f9023db54c1274ad1e72f5