Revert libp2p changes to restore QUIC [#3420]
Closes #3420. Reverts changes made in !912 that broke QUIC.

Updating to the latest version of libp2p caused the following error message when connecting to nodes:

> QUIC draft-29 has been removed, QUIC (RFC 9000) is accessible with /quic-v1

So I'm reverting to a version that does not cause these errors. I am also reverting other updates, such as those to Ethereum, because not doing so while also reverting the libp2p change lead to build errors, and the Ethereum vulnerability is not relevant to Accumulate.

Changelog: fix

Former-commit-id: 10e3161985b8735e5c6eb10b51d3d6913ab5df3c